owlmind/requirements.txt
#
# Requirements for Owlmind
# Requires Python 3.10 or higher

python-dotenv==1.0.1
requests>=2.28.0
discord.py>=2.3.2
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
typing-extensions>=4.5.0
